refactor: move trident MFA policy into ferriskey-trident
