+++ libs/ferriskey-trident/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod entities;
+pub mod mfa_policy;
